+++ src/ui/popups/scripts.rs
@@ -49,7 +49,7 @@ pub(crate) fn render(f: &mut Frame, app: &App) {
             let created = crate::ui::fmt_created(&s.modified);
             ListItem::new(Line::from(vec![
                 Span::styled(s.name.clone(), Style::default().fg(app.theme.fg)),
-                Span::styled(format!("  {}", crate::app::human_size(s.size as i64)), dim),
+                Span::styled(format!("  {}", crate::app::human_size(s.size)), dim),
                 Span::styled(format!("  {created}"), dim),
             ]))
         })
